fix(editor): stop rebuilding the document editor on every machine emission
The editable path hit the same flushSync warning as the read-only viewer, from
BaseBlockNoteView's unkeyed EditorContent. Move the editor-instance key helper
into editor-view-key.ts and apply it to both view components.

Also remove the churn feeding it. selectRenderableBlocks returned a fresh `[]`
when there was no content, and useSelector compares with Object.is — so every
machine emission looked like new blocks and recreated the whole editor. Return a
shared constant instead, and key DocumentEditor's editor on the serialized
content rather than on `initialContent`'s identity, so refetches that hand back
identical content no longer tear down every ProseMirror node view. `resourceId.id`
joins the deps so navigating between documents with identical content still
builds a new editor, since the option closures capture `resourceId`.
